Fixes, cleaned up code, added a bunch of new functions and classes, and fixed local to world space maths again...

+++ src/math/utils/utils.rs
@@ -1,23 +1,23 @@
 use rand::Rng;
-use crate::constants::PI;
+use crate::PI;
 
 #[allow(dead_code)]
 #[inline(always)]
-pub fn degrees_to_radians(degrees: f32) -> f32
+pub fn degrees_to_radians(degrees: f64) -> f64
 {
-    return degrees * (PI as f32) / 180.0;
+    return degrees * (PI as f64) / 180.0;
 }
 
 #[allow(dead_code)]
 #[inline(always)]
-pub fn random_float(min: f32, max: f32) -> f32
+pub fn random_float(min: f64, max: f64) -> f64
 {
     return rand::thread_rng().gen_range(min..max);
 }
 
 #[allow(dead_code)]
 #[inline(always)]
-pub fn clamp(x: f32, min: f32, max: f32) -> f32
+pub fn clamp(x: f64, min: f64, max: f64) -> f64
 {
     if x < min
     {
